bsps/i386: Make console output to com1 by default
RTEMS OS is most often used in embedded systems without VGA or other
port for the display connecting and sometimes it is not possible to
override the boot parameters using the OS loader, for example, if we use
a specific loader or multiboot in grub, as in the rtems-ec-cli project:

https://github.com/maxpoliak/rtems-ec-cli

when coreboot + seabios + grub is used as a loader on real hardware.

Make the console output to com1 by default (it would be better to override
this setting for output to the VGA port).

Signed-off-by: Maxim Polyakov <max.senia.poliak@gmail.com>